+++ src/claude_workspaces/ui/builders/sidebar_builder.py
@@ -499,6 +499,9 @@ class SidebarBuilder:
         self.console_runners_remove_requested = (
             self._footer.console_runners_remove_requested
         )
+        self.console_stack_raise_requested = (
+            self._footer.console_stack_raise_requested
+        )
         self.set_console_runners = self._footer.set_console_runners
         self.version_label = self._footer.version_label
         self.minimized_tray = self._footer.minimized_tray